+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * KLS-368 - Reskin styling and bug fixes
 * KLS-395 - Update django to 3.2.18, directory-sso-api-client to >= 7.1.1
 * KLS-343 Add Export academy landing page
+* KLS-404 - Add time filters to the Export Academy event listings page
 
 ## [2.25.1](https://github.com/uktrade/great-cms/releases/tag/2.25.1)